feat: Modify Button component
1. I moved the Button's dir to common dir.
2. I added a handleClick props to handle Click event.
